Move shared functions into `common.py` and delete ingress
